🚚 move deployCommands to handlers
